+++ src/main/java/com/andersmmg/lockandblock/record/KeycardReaderPacket.java
@@ -2,5 +2,5 @@ package com.andersmmg.lockandblock.record;
 
 import net.minecraft.util.math.BlockPos;
 
-public record KeycardReaderPacket(BlockPos pos, boolean remove, String uuid) {
+public record KeycardReaderPacket(BlockPos pos, KeycardReaderPacketType type, String uuid) {
 }