+++ src/main/java/com/andersmmg/lockandblock/record/KeycardReaderPacketType.java
@@ -1,0 +1,5 @@
+package com.andersmmg.lockandblock.record;
+
+public enum KeycardReaderPacketType {
+    REMOVE, CLEAR, TOGGLE_ON, TOGGLE_OFF
+}